Squashed commit of the following:
commit 7e484e08ad35bf5a3bd13cbc5d152490af80870c
Author: sanderegg <35365065+sanderegg@users.noreply.github.com>
Date:   Tue Jun 28 16:21:27 2022 +0200

    correctly handles 404

commit 261d8786c820115d9176c45ada459f2deb4680ec
Author: sanderegg <35365065+sanderegg@users.noreply.github.com>
Date:   Tue Jun 28 16:21:00 2022 +0200

    properly setup dsm middleware

commit 8ba01277095142044ec69ea17252f1af0accf3c5
Author: sanderegg <35365065+sanderegg@users.noreply.github.com>
Date:   Tue Jun 28 15:40:20 2022 +0200

    added more error handling in the handler

commit ff10ef4615157fb4fa63b8cbb72c941e62808df8
Author: sanderegg <35365065+sanderegg@users.noreply.github.com>
Date:   Tue Jun 28 14:17:59 2022 +0200

    reverted bad change

commit 3517570fba0f275f054dee64348a16776eb0a431
Author: sanderegg <35365065+sanderegg@users.noreply.github.com>
Date:   Tue Jun 28 14:17:04 2022 +0200

    separated storage and datcore

commit cd4f15e1dc354ec94484fbad4f8d24670c1ed570
Author: sanderegg <35365065+sanderegg@users.noreply.github.com>
Date:   Tue Jun 28 09:57:17 2022 +0200

    fix interface
