+++ services/storage/src/simcore_service_storage/application.py
@@ -47,13 +47,15 @@ def create(settings: Settings) -> web.Application:
         setup_db(app)  # -> postgres service
     if settings.STORAGE_S3:
         setup_s3(app)  # -> minio service
+
+    setup_rest(app)
+
     if settings.STORAGE_POSTGRES and settings.STORAGE_S3:
         setup_dsm(app)  # core subsystem. Needs s3 and db setups done
         if settings.STORAGE_CLEANER_INTERVAL_S:
             setup_dsm_cleaner(app)
-        app.middlewares.append(dsm_exception_handler)
 
-    setup_rest(app)  # lastly, we expose API to the world
+        app.middlewares.append(dsm_exception_handler)
 
     if settings.LOG_LEVEL == "DEBUG":
         setup_dev_error_logger(app)
