+++ services/storage/src/simcore_service_storage/db_file_meta_data.py
@@ -5,7 +5,7 @@ import sqlalchemy as sa
 from aiopg.sa.connection import SAConnection
 from models_library.projects import ProjectID
 from models_library.projects_nodes import NodeID
-from models_library.projects_nodes_io import StorageFileID
+from models_library.projects_nodes_io import SimcoreS3FileID
 from models_library.users import UserID
 from models_library.utils.fastapi_encoders import jsonable_encoder
 from simcore_postgres_database.models.file_meta_data import file_meta_data
@@ -16,17 +16,7 @@ from .exceptions import FileMetaDataNotFoundError
 from .models import FileMetaData, FileMetaDataAtDB
 
 
-def _convert_db_to_model(x: FileMetaDataAtDB) -> FileMetaData:
-    return FileMetaData.parse_obj(
-        x.dict()
-        | {
-            "file_uuid": x.file_id,
-            "file_name": x.file_id.split("/")[-1],
-        }
-    )
-
-
-async def fmd_exists(conn: SAConnection, file_id: StorageFileID) -> bool:
+async def fmd_exists(conn: SAConnection, file_id: SimcoreS3FileID) -> bool:
     return (
         await conn.scalar(
             sa.select([sa.func.count()])
@@ -39,20 +29,23 @@ async def fmd_exists(conn: SAConnection, file_id: StorageFileID) -> bool:
 
 async def upsert_file_metadata_for_upload(
     conn: SAConnection, fmd: FileMetaData
-) -> FileMetaData:
+) -> FileMetaDataAtDB:
     # NOTE: upsert file_meta_data, if the file already exists, we update the whole row
     # so we get the correct time stamps
     fmd_db = FileMetaDataAtDB.from_orm(fmd)
     insert_statement = pg_insert(file_meta_data).values(**jsonable_encoder(fmd_db))
     on_update_statement = insert_statement.on_conflict_do_update(
         index_elements=[file_meta_data.c.file_id], set_=jsonable_encoder(fmd_db)
-    )
-    await conn.execute(on_update_statement)
+    ).returning(literal_column("*"))
+    result = await conn.execute(on_update_statement)
+    row = await result.first()
+    assert row  # nosec
+    return FileMetaDataAtDB.from_orm(row)
 
-    return fmd
 
-
-async def insert_file_metadata(conn: SAConnection, fmd: FileMetaData) -> FileMetaData:
+async def insert_file_metadata(
+    conn: SAConnection, fmd: FileMetaData
+) -> FileMetaDataAtDB:
     fmd_db = FileMetaDataAtDB.from_orm(fmd)
     result = await conn.execute(
         file_meta_data.insert()
@@ -61,16 +54,15 @@ async def insert_file_metadata(conn: SAConnection, fmd: FileMetaData) -> FileMet
     )
     row = await result.first()
     assert row  # nosec
-    return _convert_db_to_model(FileMetaDataAtDB.from_orm(row))
+    return FileMetaDataAtDB.from_orm(row)
 
 
-async def get(conn: SAConnection, file_id: StorageFileID) -> FileMetaData:
+async def get(conn: SAConnection, file_id: SimcoreS3FileID) -> FileMetaDataAtDB:
     result = await conn.execute(
         query=sa.select([file_meta_data]).where(file_meta_data.c.file_id == file_id)
     )
-    if row := await result.fetchone():
-        fmd_at_db = FileMetaDataAtDB.from_orm(row)
-        return _convert_db_to_model(fmd_at_db)
+    if row := await result.first():
+        return FileMetaDataAtDB.from_orm(row)
     raise FileMetaDataNotFoundError(file_id=file_id)
 
 
@@ -81,7 +73,7 @@ async def list_fmds_with_partial_file_id(
     project_ids: list[ProjectID],
     file_id_prefix: Optional[str],
     partial_file_id: Optional[str],
-) -> list[FileMetaData]:
+) -> list[FileMetaDataAtDB]:
     stmt = sa.select([file_meta_data]).where(
         (
             (file_meta_data.c.user_id == f"{user_id}")
@@ -98,11 +90,7 @@ async def list_fmds_with_partial_file_id(
             else True
         )
     )
-    file_metadatas = [
-        _convert_db_to_model(FileMetaDataAtDB.from_orm(row))
-        async for row in await conn.execute(stmt)
-    ]
-    return file_metadatas
+    return [FileMetaDataAtDB.from_orm(row) async for row in await conn.execute(stmt)]
 
 
 async def list_fmds(
@@ -110,9 +98,9 @@ async def list_fmds(
     *,
     user_id: Optional[UserID] = None,
     project_ids: Optional[list[ProjectID]] = None,
-    file_ids: Optional[list[StorageFileID]] = None,
+    file_ids: Optional[list[SimcoreS3FileID]] = None,
     expired_after: Optional[datetime.datetime] = None,
-) -> list[FileMetaData]:
+) -> list[FileMetaDataAtDB]:
 
     stmt = sa.select([file_meta_data]).where(
         and_(
@@ -127,11 +115,7 @@ async def list_fmds(
         )
     )
 
-    file_metadatas = [
-        _convert_db_to_model(FileMetaDataAtDB.from_orm(row))
-        async for row in await conn.execute(stmt)
-    ]
-    return file_metadatas
+    return [FileMetaDataAtDB.from_orm(row) async for row in await conn.execute(stmt)]
 
 
 async def number_of_uploaded_fmds(conn: SAConnection) -> int:
@@ -143,16 +127,16 @@ async def number_of_uploaded_fmds(conn: SAConnection) -> int:
 
 async def list_all_uploaded_fmds(
     conn: SAConnection,
-) -> AsyncGenerator[FileMetaData, None]:
+) -> AsyncGenerator[FileMetaDataAtDB, None]:
     """returns all the theoretically valid fmds (e.g. upload_expires_at column is null)"""
     async for row in conn.execute(
         sa.select([file_meta_data]).where(file_meta_data.c.upload_expires_at == None)
     ):
         fmd_at_db = FileMetaDataAtDB.from_orm(row)
-        yield _convert_db_to_model(fmd_at_db)
+        yield fmd_at_db
 
 
-async def delete(conn: SAConnection, file_ids: list[StorageFileID]) -> None:
+async def delete(conn: SAConnection, file_ids: list[SimcoreS3FileID]) -> None:
     await conn.execute(
         file_meta_data.delete().where(file_meta_data.c.file_id.in_(file_ids))
     )
