+++ services/storage/src/simcore_service_storage/handlers_files.py
@@ -1,9 +1,9 @@
 import logging
-from typing import Optional
+from typing import Optional, cast
 
 from aiohttp import web
 from aiohttp.web import RouteTableDef
-from models_library.api_schemas_storage import FileMetaDataGet, LinkType, SoftCopyBody
+from models_library.api_schemas_storage import FileMetaDataGet, SoftCopyBody
 from models_library.utils.fastapi_encoders import jsonable_encoder
 from pydantic import AnyUrl
 from servicelib.aiohttp.requests_validation import (
@@ -12,10 +12,11 @@ from servicelib.aiohttp.requests_validation import (
     parse_request_query_parameters_as,
 )
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
+from simcore_service_storage.simcore_s3_dsm import SimcoreS3DataManager
 
 # Exclusive for simcore-s3 storage -----------------------
 from ._meta import api_vtag
-from .constants import SIMCORE_S3_ID, SIMCORE_S3_STR
+from .dsm import get_dsm_provider
 from .models import (
     CopyAsSoftLinkParams,
     FileDownloadQueryParams,
@@ -26,8 +27,6 @@ from .models import (
     LocationPathParams,
     StorageQueryParamsBase,
 )
-from .utils import get_location_from_id
-from .utils_handlers import prepare_storage_manager
 
 log = logging.getLogger(__name__)
 
@@ -44,13 +43,9 @@ async def get_files_metadata(request: web.Request):
         "received call to get_files_metadata with %s",
         f"{path_params=}, {query_params=}",
     )
-
-    dsm = await prepare_storage_manager(
-        jsonable_encoder(path_params), jsonable_encoder(query_params), request
-    )
+    dsm = get_dsm_provider(request.app).get(path_params.location_id)
     data: list[FileMetaData] = await dsm.list_files(
         user_id=query_params.user_id,
-        location=get_location_from_id(path_params.location_id),
         uuid_filter=query_params.uuid_filter,
     )
     py_data = [jsonable_encoder(FileMetaDataGet.from_orm(d)) for d in data]
@@ -69,13 +64,9 @@ async def get_file_metadata(request: web.Request):
         f"{path_params=}, {query_params=}",
     )
 
-    dsm = await prepare_storage_manager(
-        jsonable_encoder(path_params), jsonable_encoder(query_params), request
-    )
-
-    data: Optional[FileMetaData] = await dsm.list_file(
+    dsm = get_dsm_provider(request.app).get(path_params.location_id)
+    data: Optional[FileMetaData] = await dsm.get_file(
         user_id=query_params.user_id,
-        location=get_location_from_id(path_params.location_id),
         file_id=path_params.file_id,
     )
     # when no metadata is found
@@ -97,28 +88,10 @@ async def download_file(request: web.Request):
         "received call to download_file with %s",
         f"{path_params=}, {query_params=}",
     )
-
-    if (
-        path_params.location_id != SIMCORE_S3_ID
-        and query_params.link_type == LinkType.S3
-    ):
-        raise web.HTTPPreconditionFailed(
-            reason=f"Only allowed to fetch s3 link for '{SIMCORE_S3_STR}'"
-        )
-
-    dsm = await prepare_storage_manager(
-        jsonable_encoder(path_params), jsonable_encoder(query_params), request
+    dsm = get_dsm_provider(request.app).get(path_params.location_id)
+    link = await dsm.create_file_download_link(
+        query_params.user_id, path_params.file_id, query_params.link_type
     )
-
-    if get_location_from_id(path_params.location_id) == SIMCORE_S3_STR:
-        link = await dsm.download_link_s3(
-            path_params.file_id, query_params.user_id, query_params.link_type
-        )
-    else:
-        link = await dsm.download_link_datcore(
-            query_params.user_id, path_params.file_id
-        )
-
     return {"error": None, "data": {"link": link}}
 
 
@@ -132,11 +105,8 @@ async def upload_file(request: web.Request):
         f"{path_params=}, {query_params=}",
     )
 
-    dsm = await prepare_storage_manager(
-        jsonable_encoder(path_params), jsonable_encoder(query_params), request
-    )
-
-    link: AnyUrl = await dsm.create_upload_link(
+    dsm = get_dsm_provider(request.app).get(path_params.location_id)
+    link: AnyUrl = await dsm.create_file_upload_link(
         user_id=query_params.user_id,
         file_id=path_params.file_id,
         link_type=query_params.link_type,
@@ -154,10 +124,8 @@ async def abort_upload_file(request: web.Request):
         f"{path_params=}, {query_params=}",
     )
 
-    dsm = await prepare_storage_manager(
-        jsonable_encoder(path_params), jsonable_encoder(query_params), request
-    )
-    await dsm.abort_upload(path_params.file_id, query_params.user_id)
+    dsm = get_dsm_provider(request.app).get(path_params.location_id)
+    await dsm.abort_file_upload(query_params.user_id, path_params.file_id)
     return web.HTTPNoContent(content_type=MIMETYPE_APPLICATION_JSON)
 
 
@@ -170,15 +138,8 @@ async def delete_file(request: web.Request):
         f"{path_params=}, {query_params=}",
     )
 
-    dsm = await prepare_storage_manager(
-        jsonable_encoder(path_params), jsonable_encoder(query_params), request
-    )
-    await dsm.delete_file(
-        user_id=query_params.user_id,
-        location=get_location_from_id(path_params.location_id),
-        file_id=path_params.file_id,
-    )
-
+    dsm = get_dsm_provider(request.app).get(path_params.location_id)
+    await dsm.delete_file(query_params.user_id, path_params.file_id)
     return web.HTTPNoContent(content_type=MIMETYPE_APPLICATION_JSON)
 
 
@@ -192,10 +153,9 @@ async def copy_as_soft_link(request: web.Request):
         f"{path_params=}, {query_params=}, {body=}",
     )
 
-    dsm = await prepare_storage_manager(
-        params={"location_id": SIMCORE_S3_ID},
-        query=jsonable_encoder(query_params),
-        request=request,
+    dsm = cast(
+        SimcoreS3DataManager,
+        get_dsm_provider(request.app).get(SimcoreS3DataManager.get_location_id()),
     )
     file_link: FileMetaData = await dsm.create_soft_link(
         query_params.user_id, path_params.file_id, body.link_id
